# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Watch Tasks API
  description: API REST para gestão de tarefas, categorias, colaboração e relatórios.
  version: 1.0.0
servers:
  - url: https://5v6yp86rk3.execute-api.us-east-2.amazonaws.com/dev
    description: AWS API Gateway (stage dev)
  - url: http://localhost:3000
    description: Desenvolvimento local
paths:
  /health:
    get:
      summary: Health check
      tags: [Sistema]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /api/auth/register:
    post:
      summary: Criar conta
      tags: [Autenticação]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserPublic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: E-mail já cadastrado
  /api/auth/login:
    post:
      summary: Login (JWT)
      tags: [Autenticação]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token e dados do usuário
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/UserPublic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Credenciais inválidas
  /api/categories:
    get:
      summary: Listar categorias do usuário
      tags: [Categorias]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Criar categoria
      tags: [Categorias]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Nome duplicado
  /api/categories/{id}:
    get:
      summary: Obter categoria
      tags: [Categorias]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      summary: Atualizar categoria
      tags: [Categorias]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Excluir categoria
      tags: [Categorias]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Sem conteúdo
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/tasks:
    get:
      summary: Listar tarefas (próprias e como colaborador)
      tags: [Tarefas]
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Criar tarefa
      tags: [Tarefas]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/tasks/{id}:
    get:
      summary: Obter tarefa
      tags: [Tarefas]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      summary: Atualizar tarefa
      tags: [Tarefas]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Sem permissão (ex. categoria só para dono)
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Excluir tarefa (apenas dono)
      tags: [Tarefas]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '204':
          description: Sem conteúdo
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/tasks/{id}/collaborators:
    get:
      summary: Listar colaboradores (apenas dono)
      tags: [Colaboração]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collaborator'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Adicionar colaborador por e-mail (apenas dono)
      tags: [Colaboração]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Apenas dono
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/tasks/{id}/collaborators/{userId}:
    delete:
      summary: Remover colaborador (apenas dono)
      tags: [Colaboração]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Sem conteúdo
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/reports/summary:
    get:
      summary: Relatório simples (totais por status e categoria)
      tags: [Relatórios]
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TaskIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    UserPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_owner:
          type: boolean
    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        categoryId:
          type: string
          format: uuid
        status:
          type: string
          enum: [todo, in_progress, done]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        categoryId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
    Collaborator:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
    ReportSummary:
      type: object
      properties:
        total_tasks:
          type: integer
        by_status:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
        by_category:
          type: array
          items:
            type: object
            properties:
              category_name:
                type: string
              count:
                type: integer
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotFound:
      description: Não encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string